match_ports is really a behavior
instead of having a whole bunch of related callbacks on draw_graphs,
put those all into what's really a "behavior" but we have to call it a "conduct"
for now because i mistakenly gave the name "behavior" to "modes" because i thought
they wouldn't possibly mutate into Montage-style modes but they really have.

a behavior is sort of a policy and reaction to a mode? now i'm unsure