do away with secondClickEvent
it's dblclick, right? but it was triggering on a single click because
the way we do selection changed.

and it just makes more sense for the label_things mode to look out for itself and install its own
dblclick handler, instead of - wha? - draw_graphs doing it

turns out you can do a lot if modes can inspect eachother - here we expose select_things' thinginess