make brush and zoom mutually exclusive
fixes #56

But any mouse button being down should lock the current mode:
1. disable brush and enable zoom when alt key pressed, but only if mouse button is not down
2. enable brush and disable zoom when alt key released, but only if mouse button is not down
3. if mouse button is released and we did not enable brush/disable zoom because of rule #2,
do that now

This is all very specific for now (why would zoom know about brush? why is zoom still
part of dc_graph.diagram?), but the idea is to generalize it to N modes, each with their
own modifier key, in the future. That might get complicated! And this is an experiment
toward that.