claude wip: massive refactor and cleanup
we have moved too fast and we have broken too many things
but i think this is recoverable

we are linting await everywhere
we are linting no code before imports
we are in the middle of debugging webworkers
we have better error messages on many dagre worker problems

we're moving toward src/workers/*-intro.js for initialization
that should not be rolled up. with careful thought our
workers might make sense again

i just entirely broke d3v4-force - apparently everyone DOES
need to share d3.dispatch and no other. hope d3@7 helps with
this

zoom has been broken for a while

we lost updates somewhere recently, only new nodes are shown
in place, if that

we are defaulting worker=false and worker=true is only partly
debugged

we have a spurious workerName, i think

we are using display_error in all demos
it has its faults and we need a system for non-fatal errors
but at least we are catching errors